requirements.txt
aiofiles==0.8.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.17.0
async-generator==1.10
async-timeout==4.0.2
asyncio-extras==1.3.2
asyncpraw==7.1.0
asyncprawcore==1.5.1
attrs==23.1.0
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
colorthief==0.2.1
decorator==4.4.2
discord==2.2.3
discord-py-slash-command==4.2.1
discord.py==2.2.3
ffmpeg==1.4
frozenlist==1.3.3
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
lazy_loader==0.3
moviepy==1.0.3
multidict==6.0.4
mutagen==1.46.0
networkx==3.1
numpy==1.25.1
packaging==23.1
Pillow==9.5.0
prawcore==2.3.0
proglog==0.1.10
pycparser==2.21
pycryptodomex==3.18.0
PyNaCl==1.5.0
python-dotenv==1.0.0
pytube==15.0.0
PyWavelets==1.4.1
RedDownloader==4.2.0
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.1
tifffile==2023.7.18
tqdm==4.62.3
typing_extensions==4.6.3
update-checker==0.18.0
urllib3==2.0.2
websocket-client==1.5.2
websockets==11.0.3
yarl==1.9.2
youtube-dl @ git+https://github.com/ytdl-org/youtube-dl.git@b8a86dcf1aa837577178ae25357d8241ab4ba6c1
yt-dlp==2023.9.24
textwrap3==0.9.2
#The following requirements are required for the pingafy cog, but disabled because of memory problems 
#dlib==19.22.1
#face-recognition==1.3.0